fix: sync models and enums with billbee-csharp-sdk
- DeletedImages: List<int> -> List<long> for Deleted and NotFound
- Address: rename Id -> BillbeeId, remove Line3 and Comment properties
- Order: remove IsCanceled and RestfulPath, add ArchivedAt/RestoredAt/History
  and new HistoryEntry class
- Product: add BillOfMaterial, Condition, Width/Length/HeightCm, IsLowStock,
  ExportDescriptionMultiLanguage; mark ExportDescription as [Obsolete];
  add BomSubArticleApiModel class
- PaymentTypeEnum: add values 126-138 (MolliePrzelewy24 through Lena)

+++ Billbee.Net/Models/DeletedImages.cs
@@ -4,7 +4,7 @@ namespace Billbee.Net.Models
 {
     public class DeletedImages
     {
-        public List<int> Deleted { get; set; }
-        public List<int> NotFound { get; set; }
+        public List<long> Deleted { get; set; }
+        public List<long> NotFound { get; set; }
     }
 }
